fix: rename code-gen.js -> gen-code-c.js, rename generateCode -> generateC/JS
